flask/requirements.txt
torch==2.10.0
torchvision==0.25.0
Flask==3.1.2
pillow==11.3.0
huggingface_hub==0.35.0
flask-cors==5.0.0
requests==2.32.5
dotenv